refactor: move files around, remove barrel files
