--- packages/snippetz/src/core/index.ts
@@ -1,5 +1,0 @@
-export * from './utils/arrayToObject'
-export * from './utils/needsQuotes'
-export * from './utils/objectToString'
-
-export * from './types'
