+++ packages/snippetz/src/index.ts
@@ -1,2 +1,2 @@
 export * from './snippetz'
-export * from './core/types'
+export * from './types'
