+++ packages/snippetz/src/plugins/clojure/clj_http/clj_http.ts
@@ -1,5 +1,5 @@
-import type { Plugin } from '@/core'
-import { convertWithHttpSnippetLite } from '@/core/utils/convertWithHttpSnippetLite'
+import type { Plugin } from '@/types'
+import { convertWithHttpSnippetLite } from '@/utils/convertWithHttpSnippetLite'
 // @ts-expect-error no types available
 import { clj_http } from '~httpsnippet-lite/dist/esm/targets/clojure/clj_http/client.mjs'
