+++ packages/snippetz/src/plugins/java/unirest/unirest.ts
@@ -1,5 +1,5 @@
-import type { Plugin } from '@/core'
-import { convertWithHttpSnippetLite } from '@/core/utils/convertWithHttpSnippetLite'
+import type { Plugin } from '@/types'
+import { convertWithHttpSnippetLite } from '@/utils/convertWithHttpSnippetLite'
 // @ts-expect-error no types available
 import { unirest } from '~httpsnippet-lite/dist/esm/targets/java/unirest/client.mjs'
