+++ packages/snippetz/src/plugins/js/jquery/jquery.ts
@@ -1,5 +1,5 @@
-import type { Plugin } from '@/core'
-import { convertWithHttpSnippetLite } from '@/core/utils/convertWithHttpSnippetLite'
+import type { Plugin } from '@/types'
+import { convertWithHttpSnippetLite } from '@/utils/convertWithHttpSnippetLite'
 // @ts-expect-error no types available
 import { jquery } from '~httpsnippet-lite/dist/esm/targets/javascript/jquery/client.mjs'
