+++ packages/snippetz/src/plugins/js/xhr/xhr.ts
@@ -1,5 +1,5 @@
-import type { Plugin } from '@/core'
-import { convertWithHttpSnippetLite } from '@/core/utils/convertWithHttpSnippetLite'
+import type { Plugin } from '@/types'
+import { convertWithHttpSnippetLite } from '@/utils/convertWithHttpSnippetLite'
 // @ts-expect-error no types available
 import { xhr } from '~httpsnippet-lite/dist/esm/targets/javascript/xhr/client.mjs'
