+++ packages/snippetz/src/snippetz.ts
@@ -33,7 +33,7 @@ import { shellHttpie } from '@/plugins/shell/httpie'
 import { shellWget } from '@/plugins/shell/wget'
 import { swiftNsurlsession } from '@/plugins/swift/nsurlsession'
 
-import type { ClientId, Plugin, Request, TargetId } from './core'
+import type { ClientId, Plugin, Request, TargetId } from './types'
 
 /**
  * Generate code examples for HAR requests
