+++ packages/snippetz/src/utils/arrayToObject.ts
